refactor(metric-lsb): separate metric lsb tree components
